+++ tests/test.lunaris.utils.js
@@ -196,4 +196,63 @@ describe('utils', () => {
 
 
   });
+
+  describe('clone', () => {
+
+    it('should clone null', () => {
+      should(utils.clone(null)).eql(null);
+    });
+
+    it('should clone undefined', () => {
+      should(utils.clone(undefined)).eql(undefined);
+    });
+
+    it('should clone base type', () => {
+      should(utils.clone('a')).eql('a');
+      should(utils.clone(1)).eql(1);
+      should(utils.clone(1.23)).eql(1.23);
+    });
+
+    it('should clone an object', () => {
+      let obj = { label : 'A' };
+      let res = utils.clone(obj);
+      should(res).eql(obj);
+      res.label = 'B';
+      should(res).not.eql(obj);
+    });
+
+    it('should clone an object with deps', () => {
+      let obj = {
+        label    : 'A',
+        subObj   : { type : 1 },
+        subArray : [
+          { label : 'a' },
+          { label : 'b' }
+        ]
+      };
+      let res = utils.clone(obj);
+      should(res).eql(obj);
+      res.label                 = 'B';
+      res.subObj.otherAttribute = 2;
+      res.subArray[0].label     = 'c';
+      should(res).not.eql(obj);
+      should(res.subObj).not.eql(obj.subObj);
+      should(res.subArray).not.eql(obj.subArray);
+      res.subArray.push({ label : 'd' });
+      should(obj.subArray.length).eql(2);
+    });
+
+    it('should clone an array', () => {
+      let obj = [
+        { label : 'A' },
+        { label : 'B' },
+        { label : 'C' }
+      ];
+      let res = utils.clone(obj);
+      should(res).eql(obj);
+      res[1].label = 'B-1';
+      should(res).not.eql(obj);
+    });
+
+  });
 });